backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymssql==2.2.11
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
fpdf2==2.7.8